+++ CHANGELOG.md
@@ -9,7 +9,7 @@ Changelog for NeoFS Node
 
 ### Changed
 - SN caches up to 1000 bearer token verification results until the next epoch (#3369)
-- SN caches up to 1000 container session token verification results until the next epoch (#3369)
+- SN caches up to 1000 session token verification results until the next epoch (#3369)
 
 ### Removed